backend/pyproject.toml
[project]
name = "aaron2-backend"
version = "0.1.0"
description = "Aaron 2.0 — AI-powered grocery, meal planning & health tracker"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "openai>=1.50.0",
    "redis[hiredis]>=5.2.0",
    "tenacity>=9.0.0",
    "httpx>=0.27.0",
    "celery[redis]>=5.4.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "respx>=0.21.0",
    "factory-boy>=3.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

# Strict checks that work with SQLModel/SQLAlchemy
check_untyped_defs = true
warn_redundant_casts = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true

# Disabled: SQLModel column expressions (.desc(), selectinload, etc.)
# use runtime metaclass magic that mypy cannot follow with strict typing.
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_any_generics = false
warn_return_any = false
